requirements.txt
Django==5.0.10
djangorestframework==3.15.2
python-docx==1.1.2
pdfminer.six==20231228
pytest==8.3.4
pytest-django==4.9.0
